ci: Add an Ubuntu:22.04 builder for RISC-V
cpuinfo is built for riscv64 using a riscv64 container.  binfmt_misc
allows the riscv64 binaries in the container to be executed with QEMU.
This is slower than cross compiling but as there's not that much code
the build times are acceptable.  It takes just under 6 minutes for the
full riscv64 github action to run.  We also have the option of running
some of the built RISC-V binaries, e.g., unit tests, in the CI. It
should be easy to expand the matrix to add CI for other architectures
not natively supported by github actions.